requirements.txt
crewai>=0.28.0
langchain>=0.1.10
langchain-core>=0.1.27
langchain-community>=0.0.19
openai>=1.13.0
anthropic>=0.8.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
seaborn>=0.13.0
streamlit>=1.31.0
streamlit-folium>=0.15.0
folium>=0.14.0
python-dotenv>=1.0.0
requests>=2.31.0